feat: 전략 패턴 적용 - `Game` 객체의 생성자에서 자동차의 이동 전략을 받도록 변경
