+++ src/main/java/racing_car/model/Game.java
@@ -1,40 +1,31 @@
 package racing_car.model;
 
+import racing_car.strategy.MoveCarStrategy;
+
 import java.util.Collections;
-import java.util.Random;
 import java.util.stream.Collectors;
 
 public class Game {
 
     private final Cars cars;
 
-    private static final int MIN_RANGE_FOR_MOVE = 4;
+    private final MoveCarStrategy moveCarStrategy;
 
-    private static final int MAX_RANGE_FOR_MOVE = 9;
-
-    private final Random random = new Random();
-
-    public Game(Cars cars) {
+    public Game(Cars cars, MoveCarStrategy moveCarStrategy) {
         this.cars = cars;
+        this.moveCarStrategy = moveCarStrategy;
     }
 
     public Cars getCars() {
         return cars;
     }
 
-    private int generateRandomNumber() {
-        return random.nextInt(MAX_RANGE_FOR_MOVE + 1);
-    }
-
-    private void moveCar(int carIndex, int number) {
-        if (number >= MIN_RANGE_FOR_MOVE) {
-            cars.moveCar(carIndex);
-        }
-    }
-
     private void moveCar(int carIndex) {
-        int randomNumber = generateRandomNumber();
-        moveCar(carIndex, randomNumber);
+        if (!moveCarStrategy.isMovable()) {
+            return;
+        }
+
+        cars.moveCar(carIndex);
     }
 
     public void moveAllCars() {
