+++ src/main/java/racing_car/strategy/MoveCarStrategy.java
@@ -1,0 +1,6 @@
+package racing_car.strategy;
+
+public interface MoveCarStrategy {
+
+    boolean isMovable();
+}
